Extract reusable example
